requirements.txt
a2a-sdk[http-server]==0.3.25
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
